+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Other blocks are:
 - **BLS**: A filetype can now be provided to the file parser if the file extensions is ambiguous.
 - **BLS**: Ballast can now support multiple files in a single curve, for instance to produce the spectrum of mixtures.  The format is "file1 @ weight1 & file2 @ weight2" (spaces are irrelevant between symbols).
 - **MRG**: Mirage was updated to support the new representations of normal modes.
+- **MRG**: Mirage now recognizes colors specifications as sub-options of `--vid-modes`.
 - **VIZ**: Added two new representations of normal modes: as arrows are centered on the atoms (midarrows), as dual arrows showing the "positive" and "negative" displacements (dualarrows).
 - **LIB**: New module `base.aliases` to provide common aliases for keywords (e.g., spectroscopy, level of theory) regularly used within ESTAMPES.
 - **LIB**: The `Spectrum` class can now support multiple datasets and relative weights (for now purely numerical).
@@ -45,6 +46,8 @@ Other blocks are:
 - **LIB**: New alias `COLOR_NAMES` containing all RGB color dictionaries dictionaries in `data.colors`.
 - **LIB**: New conversion function from a color specification to RGB tuple, `data.colors.to_rgb_list`.
 - **LIB**: Default colors of vibrational modes are now stored in `data.visuals.VIBCOLS`.
+- **LIB**: `visual.povrender.write_pov_vib` now fully supports color specifications.
+- **LIB**: `visual.vibview.VidMode` now fully supports color specifications.
 - **API**: Multi-jobs (explicit or internal linked jobs) in Gaussian log files are now supported.  ESTAMPES assumes that the jobs are all related to the same task (for instance opt+freq) and does not support selective extractions from a specific job (e.g., linked freq jobs with different levels of theory) for now.
 - **API**: Added support of anharmonic X matrix from Gaussian log files.